Commit Fix name class Bai tap 1 Phan Drawing
