+++ 06_Drawing/AnhPD/BaiTap1/BaiTap1/ChartViewController.swift
@@ -1,0 +1,32 @@
+//
+//  ChartViewController.swift
+//  BaiTap1
+//
+//  Created by PCI0012 on 4/9/20.
+//  Copyright © 2020 PCI0012. All rights reserved.
+//
+
+import UIKit
+
+class ChartViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        let paint = PaintView()
+        paint.draw()
+        view.addSubview(paint)
+        // Do any additional setup after loading the view.
+    }
+
+
+    /*
+    // MARK: - Navigation
+
+    // In a storyboard-based application, you will often want to do a little preparation before navigation
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+        // Get the new view controller using segue.destination.
+        // Pass the selected object to the new view controller.
+    }
+    */
+
+}
