+++ 06_Drawing/AnhPD/BaiTap1/BaiTap1/SceneDelegate.swift
@@ -17,7 +17,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
         window?.makeKeyAndVisible()
-        let viewController = ViewController()
+        let viewController = ChartViewController()
         window?.rootViewController = viewController
     }
