--- 06_Drawing/AnhPD/BaiTap1/BaiTap1/ViewController.swift
@@ -1,12 +1,0 @@
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        let paint = PaintView()
-        paint.draw()
-        view.addSubview(paint)
-    }
-}
